# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservasi Booking Service API",
        "description": "API untuk Service Reservasi & Booking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/v1/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get all reservations",
                "operationId": "019b99eb90234f732ab2477474963cdb",
                "responses": {
                    "200": {
                        "description": "Reservations retrieved successfully"
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get reservasi by ID",
                "operationId": "cc6de5b1aa54a04970d4a3d58989b1e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservasi found"
                    },
                    "401": {
                        "description": "Invalid API Key"
                    },
                    "404": {
                        "description": "Reservasi not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{id}/checkin": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Check In Reservasi",
                "operationId": "586caab5ecbea1e4ff6d57dd39640dbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check-in success"
                    },
                    "401": {
                        "description": "Invalid API Key"
                    },
                    "404": {
                        "description": "Reservasi not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{id}/status": {
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update reservasi status",
                "operationId": "be6e8d573e5bac70070f40dd5f87a829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "401": {
                        "description": "Invalid API Key"
                    },
                    "404": {
                        "description": "Reservasi not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Reservations",
            "description": "Reservasi API"
        }
    ]
}